third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-aspect-ratio-expected.txt

This is a testharness.js-based test.
[FAIL] Loaded images test: Error image with width, height and alt attributes
  assert_equals: expected "auto 100 / 500" but got "auto"
Harness: the test ran to completion.
